requirements.txt
# Project requirements
fastapi
celery[sqlalchemy]
sqlalchemy
pydantic
psycopg2
uvicorn
python-multipart
boto3

# YOLO requirements
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
pandas
seaborn